third_party/blink/web_tests/wpt_internal/script_tools/unregister_invalid_tool.https.html

<!DOCTYPE html>
<html>
<title>Script Tools: invalid use of register/unregister tools</title>
<link rel="author" href="mailto:khushalsagar@chromium.org">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
promise_test(async () => {
  const modelContext = navigator.modelContext;
  assert_throws_dom(
    'InvalidStateError',
    () => {
      modelContext.unregisterTool("empty");
    },
    "unregisterTool which doesn't exist"
  );
}, "unregisterTool which doesn't exist");
</script>
